pyproject.toml
[tool.poetry]
name = "glyphworks"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "n"

[tool.poetry.dependencies]
python = "^3.10.7"
django-extensions = "*"
django-webpack-loader = "*"
daphne = "*"
gunicorn = "*"
flake8 = "*"
psycopg2 = "*"
factory-boy = "*"
ipython = "*"
graphene-django = "*"
django-csp = "*"
django = "^4.1.6"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pip-audit = "*"
watchfiles = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
